chore: cleanup and remove unnecessary methods
