+++ gax/src/main/java/com/google/api/gax/retrying/DirectRetryingExecutor.java
@@ -125,9 +125,4 @@ public class DirectRetryingExecutor<ResponseT> implements RetryingExecutorWithCo
       Thread.sleep(delay.toMillis());
     }
   }
-
-  /** Returns the retry algorithm that is used by this executor. */
-  protected RetryAlgorithm<ResponseT> getRetryAlgorithm() {
-    return retryAlgorithm;
-  }
 }
