+++ gax/src/test/java/com/google/api/gax/retrying/FailingCallable.java
@@ -43,7 +43,6 @@ class FailingCallable implements Callable<String> {
           .setRetryDelayMultiplier(1)
           .setMaxRetryDelay(Duration.ofMillis(8L))
           .setInitialRpcTimeout(Duration.ofMillis(8L))
-          .setJittered(false)
           .setRpcTimeoutMultiplier(1)
           .setMaxRpcTimeout(Duration.ofMillis(8L))
           .setTotalTimeout(Duration.ofMillis(400L))
@@ -55,7 +54,6 @@ class FailingCallable implements Callable<String> {
           .setRetryDelayMultiplier(1)
           .setMaxRetryDelay(Duration.ofNanos(1L))
           .setInitialRpcTimeout(Duration.ofNanos(1L))
-          .setJittered(false)
           .setRpcTimeoutMultiplier(1)
           .setMaxRpcTimeout(Duration.ofNanos(1L))
           .setTotalTimeout(Duration.ofNanos(1L))
